requirements.txt
boto3
pandas
scikit-learn
pyyaml
